apim-for-ai-demo/guardrails-service/requirements.txt
--index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://pypi.org/simple

wheel==0.45.1

# Scientific computing
tqdm==4.67.1
packaging==25.0

# Core ML libraries (with CUDA support)
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
transformers==4.55.2
accelerate==1.10.0

# PEFT and quantization
peft==0.17.0
bitsandbytes==0.47.0

# Guardrails
guardrails-ai==0.6.6

# Build tools
ninja==1.13.0

# Template engine
jinja2==3.1.6

# Web server dependencies
fastapi==0.116.1
uvicorn==0.35.0
gunicorn==23.0.0

# PII Detection and NLP
presidio-anonymizer==2.2.359
presidio-analyzer==2.2.359
gliner==0.2.21
spacy>=3.7.0

#pip install flash-attn==2.8.3 --use-pep517 --no-build-isolation
